+++ EnhanceQoL/Locales/ruRU.lua
@@ -142,6 +142,12 @@ L["CVarOptions"] = {
 	["UnitNamePlayerGuild"] = { trueValue = "1", falseValue = "0", description = "Показывать гильдию игроков" },
 	["UnitNamePlayerPVPTitle"] = { trueValue = "1", falseValue = "0", description = "Показывать титул игроков" },
 	["WholeChatWindowClickable"] = { description = "Сделать весь чат кликабельным", trueValue = "1", falseValue = "0" },
+	["addonProfilerEnabled"] = {
+		description = "Включить профилирование аддонов от Blizzard (требует много ресурсов ЦП)",
+		trueValue = "1",
+		falseValue = "0",
+		register = true,
+	},
 }
 
 L["autoAcceptGroupInvite"] = "Автоматически принимать приглашения в группу"