+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Added
 
 - **Mythic Plus**
+
   - **Talent Reminder**
     - Add an option to set your prefered Talent Loadout for each spec and Mythic Plus Dungeon to be reminded when it's wrong
+
+- **New CVar Option**
+  - **Blizzard Addon Profiling**
+    - Option to disable the Blizzard Addon Profiling, which has some heavy CPU-Load